Implement world creation.
In order to generate a proper level.dat per world, every world will now
use a seperate save handler unlike Forge/Vanilla which use only one.
This is required in order to have world settings saved per world.

In order to create a custom world, you will be required to use our new
WorldBuilder. This allows you to create and load a world immediately or
simply generate a world for later use.

Note: When creating a world, it will NOT be loaded. The world folder is
created along with initial data. If you want the world to be available,
you must load it by calling loadWorld.

After a world is created, Sponge will register these worlds,
if enabled, during server startup so they can be loaded automatically with
the default worlds. This should help plugins such as MultiVerse which had
to support this themselves.

UUID support has been added and is now stored with the world data.

All custom world data will be located in a new level dat file called
"level_sponge.dat". We chose to store all additional world data in a
seperate file to avoid loss of data if a user launched the server in a
non-forge or sponge environment.